+++ Headerfile/VoltageCalculation.h
@@ -10,8 +10,8 @@
 #define UNDER_VOLTAGE_THRESHOLD 28.0 // Threshold for undervoltage in volts
 
 // Voltage divider resistances
-extern const double R1; // Resistance of R1 in ohms
-extern const double R2; // Resistance of R2 in ohms
+extern const uint32_t R1; // Resistance of R1 in ohms
+extern const uint32_t R2; // Resistance of R2 in ohms
 
 /**
  * @brief Calculates the input voltage based on the ADC value.
@@ -23,6 +23,6 @@ extern const double R2; // Resistance of R2 in ohms
  * @param adc_value The ADC value to convert to voltage.
  * @return The calculated input voltage.
  */
-double calculate_voltage(unsigned int adc_value);
+uint32_t calculate_voltage(uint32_t adc_value);
 
 #endif // VOLTAGE_CALCULATION_H
